+++ integration-tests/allow-apply-all/hive.nix
@@ -1,0 +1,14 @@
+let
+  tools = import ./tools.nix {
+    insideVm = true;
+    targets = ["alpha"];
+  };
+in {
+  meta = {
+    nixpkgs = tools.pkgs;
+    allowApplyAll = false;
+  };
+
+  deployer = tools.getStandaloneConfigFor "deployer";
+  alpha = tools.getStandaloneConfigFor "alpha";
+}